developed delete function
